Fix the issue by modifying the logic of deleting the page for retake feature.
The root reason caused this issue is that the current page is deleted when clicking the "Retake" button, causing the logic of the adding page to be incorrect and an infinite loop. Fix the issue by deleting the current page when the new retaken picture is taken.

CC-4829 AND - ANR in Documents scanner