Fix the issue by displaying the filename using a TextView once the editing process is finished.
Because the legacy code is using the customised view - SuffixTextInputEditText to show the file name and it's difficult to solve the displayed issue by still using this view, fix the issue by displaying the filename using a TextView once the editing process is finished.

Also, optimised the behaviour to improve the user experience.
The existing behaviour:
1. If the edited filename is too long, the filename is hidden middle part to display suffixes for users.
2. When the file properties are changed by users during the filename editing, the edit is finished and shows the filename with a suffix.
3. If the filename is null or invalid, and users attempt to complete the edit, a notification message is presented to inform the users.
4. If the filename is null or invalid, and users click the back button, the filename is reverted to the previous filename

CC-5415 AND - file extension cutoff when a long file name in save scan